requirements.txt
Django==2.1.7
djangorestframework==3.9.2
markdown==3.0.1
django-filter==2.1.0
coreapi==2.3.3
djangorestframework_simplejwt==4.1.0
zxcvbn-python==4.4.24
django-rest-swagger==2.2.0
ldap3==2.6